refactor/first step to refactor application to hexagonal architecture. Create domains object, commands, handlers, validator etc.
