+++ src/Application/Command/AssignDelegationToUser.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace App\Application\Command;
+
+final class AssignDelegationToUser
+{
+ public readonly int $delegationId;
+
+ public readonly int $userId;
+}
