+++ src/Application/Command/CreateDelegation.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace App\Application\Command;
+
+class CreateDelegation
+{
+    public string $country;
+
+    public \DateTimeInterface $startDelegation;
+
+    public \DateTimeInterface $endDelegation;
+
+    public \DateTimeInterface $createdAt;
+}
