+++ src/Application/Command/CreateUser.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace App\Application\Command;
+
+class CreateUser
+{
+  public string $id;
+}
