+++ src/Application/Query/DelegationQuery.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace App\Application\Query;
+
+interface DelegationQuery
+{
+
+}
