+++ src/Application/Query/UserQuery.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace App\Application\Query;
+
+interface UserQuery
+{
+
+}
