+++ src/Domain/Delegation/CalculatorStrategy/ZeroCostDelegation.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Domain\Delegation\CalculatorStrategy;
+
+use App\Domain\Delegation\CostDelegationInterface;
+
+final class ZeroCostDelegation implements CostDelegationInterface
+{
+    public function calculate(int $numberOfDaysOnDelegation): int
+    {
+        return 0 * $numberOfDaysOnDelegation;
+    }
+}
