+++ src/Domain/Delegation/CostDelegationInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Domain\Delegation;
+
+interface CostDelegationInterface
+{
+    public function calculate(int $numberOfDaysOnDelegation): int;
+}
