+++ src/Domain/User/UseRepository.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace App\Domain\User;
+
+interface UseRepository
+{
+    public function save(User $user): void;
+}
