+++ src/Domain/User/User.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Domain\User;
+
+class User
+{
+    private ?int $id;
+
+    public function getId(): ?int
+    {
+        return $this->id;
+    }
+}
