--- src/Enum/Country.php
@@ -1,12 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Enum;
-
-enum Country: int
-{
-    case PL = 10;
-    case DE = 50;
-    case GB = 75;
-}
