+++ src/Infrastructure/ORM/Mapping/Delegation.orm.xml
@@ -1,0 +1,26 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<doctrine-mapping xmlns="http://doctrine-project.org/schemas/orm/doctrine-mapping"
+                  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+                  xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping https://www.doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
+
+    <entity name="App\Domain\Delegation\Delegation" table="delegations">
+
+        <id name="id" type="integer" column="id">
+            <generator/>
+        </id>
+
+        <field name="country" column="country"/>
+
+        <field name="startDelegation" column="start_delegation" type="datetime"/>
+
+        <field name="endDelegation" column="end_delegation" type="datetime"/>
+
+        <field name="createdAt" column="created_at" type="datetime"/>
+
+        <many-to-one field="user" target-entity="User">
+            <join-column name="user_id"/>
+        </many-to-one>
+
+    </entity>
+
+</doctrine-mapping>
